backend/requirements.txt
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.4
psycopg2-binary==2.9.11
pypdf==6.10.1
python-decouple==3.8
python-docx==1.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
